update document and shipment js

+++ www/public_html/js/document.js
@@ -561,6 +561,7 @@ $(document).ready(function () {
 
         // load the url and show modal on success
         preloader(url);
+        $('#document_action, #go_back').toggle();
     });
 
     // Button Comment
@@ -571,6 +572,7 @@ $(document).ready(function () {
         var url = "/document/comment/" + doc_id + "/write/" + doc_status;
         // load the url and show modal on successs
         preloader(url);
+        $('#document_action, #go_back').toggle();
     });
 
     // Button Edit
@@ -585,6 +587,7 @@ $(document).ready(function () {
 
         // load the url and show modal on success
         preloader(url);
+        $('#document_action, #go_back').toggle();
     });
 
     // Button Download
@@ -667,15 +670,15 @@ $(document).ready(function () {
         preloader(url);
     });
 
-    // Button Approve All
-    $('button#approve_all').click(function (ev) {
-        updateDocumentStatus("status", "approved");
-    });
+    // // Button Approve All
+    // $('button#approve_all').click(function (ev) {
+    //     updateDocumentStatus("status", "approved");
+    // });
 
-        // 
-    $('button#pending_all').click(function (ev) {
-        updateDocumentStatus("status", "pending");
-    });
+    // // Button Pending All 
+    // $('button#pending_all').click(function (ev) {
+    //     updateDocumentStatus("status", "pending");
+    // });
 
     // Custom: Add checkbox in each file - preview - frame
     $('.file-preview-thumbnails > .file-preview-frame').each(function () {
@@ -689,6 +692,7 @@ $(document).ready(function () {
 
     });
 
+    // Select All Checkbox
     $('#select-all').click(function() {
         var checked = false;
         if(this.checked) {
@@ -845,9 +849,16 @@ $(document).ready(function () {
         }
     });
 
+    // Do action in selected documents
+    $("#push_selected, #delete_selected, #approve_all, #approve_selected, #pending_all, #pending_selected ").on("click", function() {
+        updateDocumentStatus($(this).data('option'), $(this).data('action'), $(this).data('text'));
+    });
+
+    // 
     $('.file-preview-frame').on('click', function(ev) {
         var checkBoxes = $(this).find('.checkbox > input');
         checkBoxes.prop("checked", !checkBoxes.prop("checked"));
+        checkBoxes.parent().toggleClass('selected');
     });
 
     // Other functions
@@ -963,53 +974,121 @@ $(function () {
     });
 });
 
-function updateDocumentStatus(groupSelected, valueSelected) {
+// updateDocumentStatus("status", "pending");
+function updateDocumentStatus(option, action, text) {
     var data = [];
-    $('.file-preview-thumbnails > .file-preview-frame > .checkbox').each(function(index, element) {
-        data.push($(element).find('input').val());
-    });
-    var msg = 'This will change the status of the document to "' + valueSelected.charAt(0).toUpperCase() + valueSelected.slice(1) + '". Do you still want to continue?';
+    var msg = "";
+    if(option == "status_all") {
+        $('.file-preview-thumbnails > .file-preview-frame > .checkbox').each(function(index, element) {
+            data.push($(element).find('input').val());
+        });
+        warning_msg = "This will <b>" + text + "</b> the document. Do you still want to continue?";
+        success_msg = "Successfully <b>" + text + "</b> the document!";
+    } else {
+        $('div[class*="selected"] > input').each(function () {
+            data.push($(this).val().replace('d-', ''));
+        });
+        warning_msg = "This action will <b>" + text + "</b> the selected document. <br>Do you still want to continue?";
+        success_msg =  "Successfully <b>" + text + "</b> the selected document!";
+    }
     Swal.fire({
-        title: 'Are you sure?',
-        text: msg,
-        icon: 'warning',
-        showCancelButton: true,
+        title: "Are you sure?",
+        html: warning_msg,
+        icon: "warning",
         showLoaderOnConfirm: true,
-        // reverseButtons: true,
+        confirmButtonText: 'Yes, continue!',
+        showCancelButton: true,
+        cancelButtonText: 'No, cancel!',
+        reverseButtons: true,
         preConfirm: () => {
-            return $.post("/document/updateDocumentBulk", { group: groupSelected, value: valueSelected, data: data })
+            return $.post("/document/updateDocumentBulk", { group: option, value: action, data: data })
             .done(function (res) {
-                data.forEach(function (entry) {
-                    var d = $('.d-' + entry);
-                    if (groupSelected === 'status') {
-                        $('[data-key="' + entry + '"]').attr("data-doc_status", valueSelected);
-                        d.find('.file-footer-caption > #status').text(valueSelected);
-                        if(valueSelected === "approved") {
-                            d.find('.file-footer-caption > #status').removeClass('pending');
-                            d.find('.file-footer-caption > #status').addClass('approved');
-                            d.removeClass("b-pending");
-                            d.addClass("b-approved");
-                            d.find('.kv-file-status').children().removeClass('pending');
-                            d.find('.kv-file-status').children().addClass('approved');
-                            d.find('.kv-file-status').children().removeClass('fa-thumbs-down');
-                            d.find('.kv-file-status').children().addClass('fa-thumbs-up');
-                            d.addClass('bg-approved');
-                            d.removeClass('bg-pending');
-                        } else {
-                            d.find('.file-footer-caption > #status').removeClass('approved');
-                            d.find('.file-footer-caption > #status').addClass('pending');
-                            d.removeClass("b-approved");
-                            d.addClass("b-pending");
-                            d.find('.kv-file-status').children().removeClass('approved');
-                            d.find('.kv-file-status').children().addClass('pending');
-                            d.find('.kv-file-status').children().removeClass('fa-thumbs-up');
-                            d.find('.kv-file-status').children().addClass('fa-thumbs-down');
-                            d.addClass('bg-pending');
-                            d.removeClass('bg-approved');
-                        }
+                if(res == "false") {
+                    if (option == 'status') {
+                        Swal.fire({
+                            title: 'It is better to leave a comment',
+                            text: 'Do you want to leave a comment?',
+                            icon: 'info',
+                            showDenyButton: true,
+                            showCancelButton: true,
+                            confirmButtonText: `Leave comment`,
+                            denyButtonText: `No`,
+                        }).then((result) => {
+                            /* Read more about isConfirmed, isDenied below */
+                            if (result.isConfirmed) {
+                                Swal.fire('Status changed!', 'You will be redirected to comment form.', 'success');
+                                var url = "/document/comment/" + data + "/write/" + action;
+                                // load the url and show modal on success
+                                preloader(url);
+                            // } else if (result.isDenied) {
+                            //     Swal.fire('Status changed!', 'Save but no comment made.', 'info')
+                            //     $.post("/document/putDocumentComment", {
+                            //         title: "",
+                            //         message: "",
+                            //         status: action,
+                            //         user_id: user_id,
+                            //         document_id: data
+                            //     });
+                            $('#document_action, #go_back').toggle();
+                            } else {
+                                Swal.fire('Status changed!', 'Save but no comment made.', 'info')
+                                $.post("/document/putDocumentComment", {
+                                    title: "",
+                                    message: "",
+                                    status: action,
+                                    user_id: user_id,
+                                    document_id: data
+                                });
+                            }
+                        });
                     }
-                });
-                sessionStorage.setItem("changeTriggered", "1");
+                    // update document preview
+                    data.forEach(function (entry) {
+                        var d = $('.d-' + entry);
+                        if (option === 'status') {
+                            $('[data-key="' + entry + '"]').attr("data-doc_status", action);
+                            d.find('.file-footer-caption > #status').text(action);
+                            if(action === "approved") {
+                                d.find('.file-footer-caption > #status').removeClass('pending');
+                                d.find('.file-footer-caption > #status').addClass('approved');
+                                d.removeClass("b-pending");
+                                d.addClass("b-approved");
+                                d.find('.kv-file-status').children().removeClass('pending');
+                                d.find('.kv-file-status').children().addClass('approved');
+                                d.find('.kv-file-status').children().removeClass('fa-thumbs-down');
+                                d.find('.kv-file-status').children().addClass('fa-thumbs-up');
+                                d.addClass('bg-approved');
+                                d.removeClass('bg-pending');
+                            } else {
+                                d.find('.file-footer-caption > #status').removeClass('approved');
+                                d.find('.file-footer-caption > #status').addClass('pending');
+                                d.removeClass("b-approved");
+                                d.addClass("b-pending");
+                                d.find('.kv-file-status').children().removeClass('approved');
+                                d.find('.kv-file-status').children().addClass('pending');
+                                d.find('.kv-file-status').children().removeClass('fa-thumbs-up');
+                                d.find('.kv-file-status').children().addClass('fa-thumbs-down');
+                                d.addClass('bg-pending');
+                                d.removeClass('bg-approved');
+                            }
+                        } else {
+                            switch (action) {
+                                case 'deleted':
+                                    d.fadeOut(3000, function () { $(this).remove(); });
+                                    break;
+                                case 'push':
+                                    d.find('.kv-file-push').remove();
+                                    break;
+                                default:
+                                    console.log('Error: value = ' + action);
+                                    break;
+                            }
+                        }
+                    });
+                    $('#select-all').prop('checked', false);
+                    sessionStorage.setItem("changeTriggered", "1");
+                }
+                return { response: res }
             }).fail(function (response) {
                 console.log('Error: ' + response.responseText);
             });
@@ -1017,14 +1096,19 @@ function updateDocumentStatus(groupSelected, valueSelected) {
         allowOutsideClick: () => !Swal.isLoading()
     }).then((result) => {
         if (result.isConfirmed) {
-            Swal.fire({
-                position: 'top-end',
-                icon: 'success',
-                title: 'All document was set to ' + valueSelected.charAt(0).toUpperCase() + valueSelected.slice(1),
-                showConfirmButton: false,
-                timer: 1500
-            });
-            sessionStorage.setItem("changeTriggered", "1");
+            if(result.value.response == "false") {
+                Swal.fire({
+                    // position: 'top-end',
+                    title: "Success",
+                    icon: 'success',
+                    title: success_msg,
+                    showConfirmButton: false,
+                    timer: 1500
+                });
+                sessionStorage.setItem("changeTriggered", "1");
+            } else {
+                console.log('Error: ' + result.value.response);
+            }
         } else if (
             /* Read more about handling dismissals below */
             result.dismiss === Swal.DismissReason.cancel
@@ -1036,4 +1120,10 @@ function updateDocumentStatus(groupSelected, valueSelected) {
             )
         }
     });
+}
+
+function goBack() {
+    var url = "/shipment/document/" + shipment_id + "/" + document_type;
+    preloader(url);
+    $('#document_action, #go_back').toggle();
 }
